refactor: Alert 생성 메서드를 UIViewController 익스텐션으로 이동
