+++ Diary/Controllers/DiaryListViewController.swift
@@ -53,24 +53,7 @@ final class DiaryListViewController: UIViewController {
     
     // MARK: - Methods
     
-    private func presentErrorAlert(_ error: (Error)) {
-        let errorAlert = UIAlertController(
-            title: AlertMessage.errorAlertTitle,
-            message: error.localizedDescription,
-            preferredStyle: .alert
-        )
         
-        let confirmAction = UIAlertAction(
-            title: AlertMessage.confirmActionTitle,
-            style: .default
-        )
-        
-        errorAlert.addAction(confirmAction)
-        
-        present(
-            errorAlert,
-            animated: true
-        )
     }
     
     private func configureNavigationItems() {
